temporarily scrubbed the dynamic reloading because it was unrelieable
To clarify. I noticed that the watcher would add only 0 comments over
the course of multiple hours. I think this was caused by a deleted
comment (and the fact that reddit probably invalidated its id causing
all queries to it to return empty lists)

I could have tried to detect this by having a no_change counter and
restarting with limit=None after it hits a certain threshhold but this
isn't the cleanest solution either.